[ feat ] 커뮤니티 상세 글 및 댓글 적용
